refactor fingerprintcss into a function
