+++ lib/__tests__/build-files.test.js
@@ -13,8 +13,7 @@ const glob = require('glob-promise');
 const rimraf = require('rimraf');
 const shell = require('shelljs');
 const path = require('path');
-const join = path.join;
-const crypto = require('crypto');
+const {fingerprintCss} = require('../server/utils.js');
 
 const CWD = process.cwd();
 
@@ -91,22 +90,9 @@ test('Generated table of contents', function() {
 });
 
 test('Concatenated CSS files', function() {
-  // create hash based on css content from static folder
-  let cssForHashing = fs.readFileSync(
-    path.join(__dirname, '../static/css', 'main.css')
+  const secureHash = fingerprintCss(
+    path.join(CWD, 'website', 'static', 'css', '*.css')
   );
-  const userCssFolder = glob.sync(
-    join(CWD, 'website', 'static', 'css', '*.css')
-  );
-  userCssFolder.forEach(file => {
-    const normalizedFile = path.normalize(file);
-    cssForHashing += fs.readFileSync(normalizedFile, 'utf8');
-  });
-  const secureHash = crypto
-    .createHash('sha1')
-    .update(cssForHashing, 'utf8')
-    .digest('hex');
-
   return Promise.all([
     glob(staticCSSDir + '/*.css'),
     fs.readFile(
