+++ lib/server/generate.js
@@ -37,23 +37,10 @@ async function execute() {
   const imageminOptipng = require('imagemin-optipng');
   const imageminSvgo = require('imagemin-svgo');
   const imageminGifsicle = require('imagemin-gifsicle');
-  const crypto = require('crypto');
+  const {fingerprintCss} = require('./utils.js');
+  const secureHash = fingerprintCss(join(CWD, 'static', 'css', '*.css'));
 
   commander.option('--skip-image-compression').parse(process.argv);
-
-  // create hash based on css content from static folder
-  let cssForHashing = fs.readFileSync(
-    path.join(__dirname, '../static/css', 'main.css')
-  );
-  const userCssFolder = glob.sync(join(CWD, 'static', 'css', '*.css'));
-  userCssFolder.forEach(file => {
-    const normalizedFile = path.normalize(file);
-    cssForHashing += fs.readFileSync(normalizedFile, 'utf8');
-  });
-  const secureHash = crypto
-    .createHash('sha1')
-    .update(cssForHashing, 'utf8')
-    .digest('hex');
 
   // create the folder path for a file if it does not exist, then write the file
   function writeFileAndCreateFolder(file, content) {
