+++ lib/server/utils.js
@@ -6,7 +6,9 @@
  */
 const path = require('path');
 const escapeStringRegexp = require('escape-string-regexp');
-const env = require('./env.js');
+const fs = require('fs-extra');
+const glob = require('glob');
+const crypto = require('crypto');
 
 // Return the subdirectory path from a reference directory
 // Example:
@@ -29,6 +31,7 @@ function getLanguage(file, refDir) {
   const match = regexSubFolder.exec(file);
 
   // Avoid misinterpreting subdirectory as language
+  const env = require('./env.js');
   if (match && env.translation.enabled) {
     const enabledLanguages = env.translation
       .enabledLanguages()
@@ -40,7 +43,23 @@ function getLanguage(file, refDir) {
   return null;
 }
 
+// create hash based on css content from static folder
+function fingerprintCss(cssPath) {
+  let cssForHashing = fs.readFileSync(
+    path.join(__dirname, '../static/css', 'main.css')
+  );
+  glob.sync(cssPath).forEach(file => {
+    const normalizedFile = path.normalize(file);
+    cssForHashing += fs.readFileSync(normalizedFile, 'utf8');
+  });
+  return crypto
+    .createHash('sha1')
+    .update(cssForHashing, 'utf8')
+    .digest('hex');
+}
+
 module.exports = {
   getSubDir,
   getLanguage,
+  fingerprintCss,
 };
